Pocket_Pallet/PP_MVP/backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation & Processing
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# File Processing
pandas==2.1.3
pyarrow==14.0.1
openpyxl==3.1.2
chardet==5.2.0

# Fuzzy Matching
python-Levenshtein==0.23.0
fuzzywuzzy==0.18.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Storage
boto3==1.29.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Utilities
unidecode==1.3.7
iso4217==1.11.20231117

